ADFA-936: Consume IDE logs in the ViewModel to avoid replay duplicates
GlobalBufferAppender replays its buffer on every registerConsumer().
With the fragment as consumer, each tab switch (view recreation)
re-registered and re-submitted the replayed lines into the retained
LogBuffer, duplicating history. The activity-scoped IDELogsViewModel is
now the consumer, so the replay happens once per ViewModel lifetime.

Found during on-device verification (line duplicated after switching
tabs away and back).
